Sprint 10 Цвета в картинках заменены на @color/... data class Track вынесен в отдельный файл
