+++ app/src/main/java/com/example/playlistmaker/track/Track.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.track
+
+data class Track(
+    val trackName: String,
+    val artistName: String,
+    val trackTime: String,
+    val artworkUrl100: String
+)
