+++ app/src/main/res/drawable/ic_arrow_back.xml
@@ -8,6 +8,6 @@
         android:pathData="M0,0h24v24h-24z"/>
     <path
         android:pathData="M20,11H7.83L13.42,5.41L12,4L4,12L12,20L13.41,18.59L7.83,13H20V11Z"
-        android:fillColor="#1A1B22"/>
+        android:fillColor="@color/black"/>
   </group>
 </vector>
