+++ app/src/main/res/drawable/ic_media.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="20">
   <path
       android:pathData="M18,0H6C4.9,0 4,0.9 4,2V14C4,15.1 4.9,16 6,16H18C19.1,16 20,15.1 20,14V2C20,0.9 19.1,0 18,0ZM16,5H13V10.5C13,11.88 11.88,13 10.5,13C9.12,13 8,11.88 8,10.5C8,9.12 9.12,8 10.5,8C11.07,8 11.58,8.19 12,8.51V3H16V5ZM2,4H0V18C0,19.1 0.9,20 2,20H16V18H2V4Z"
-      android:fillColor="#1A1B22"/>
+      android:fillColor="@color/black"/>
 </vector>
