+++ app/src/main/res/drawable/ic_search_small.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="14">
   <path
       android:pathData="M10.006,8.805H9.373L9.149,8.589C9.934,7.676 10.406,6.492 10.406,5.203C10.406,2.329 8.077,0 5.203,0C2.329,0 0,2.329 0,5.203C0,8.077 2.329,10.406 5.203,10.406C6.492,10.406 7.676,9.934 8.589,9.149L8.805,9.373V10.006L12.807,14L14,12.807L10.006,8.805ZM5.203,8.805C3.21,8.805 1.601,7.196 1.601,5.203C1.601,3.21 3.21,1.601 5.203,1.601C7.196,1.601 8.805,3.21 8.805,5.203C8.805,7.196 7.196,8.805 5.203,8.805Z"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
 </vector>
