+++ app/src/main/res/drawable/ic_separator.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="13">
   <path
       android:pathData="M6.5,6.5m-1.5,0a1.5,1.5 0,1 1,3 0a1.5,1.5 0,1 1,-3 0"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
 </vector>
