+++ app/src/main/res/drawable/ic_share.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="18">
   <path
       android:pathData="M13.333,12.723C12.658,12.723 12.053,12.994 11.591,13.419L5.253,9.669C5.298,9.461 5.333,9.253 5.333,9.036C5.333,8.819 5.298,8.611 5.253,8.404L11.52,4.69C12,5.142 12.631,5.422 13.333,5.422C14.809,5.422 16,4.211 16,2.711C16,1.211 14.809,0 13.333,0C11.858,0 10.667,1.211 10.667,2.711C10.667,2.928 10.702,3.136 10.747,3.343L4.48,7.057C4,6.605 3.369,6.325 2.667,6.325C1.191,6.325 0,7.536 0,9.036C0,10.536 1.191,11.747 2.667,11.747C3.369,11.747 4,11.467 4.48,11.015L10.809,14.774C10.764,14.964 10.738,15.163 10.738,15.361C10.738,16.816 11.902,18 13.333,18C14.764,18 15.929,16.816 15.929,15.361C15.929,13.907 14.764,12.723 13.333,12.723Z"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
 </vector>
