+++ app/src/main/res/drawable/ic_support.xml
@@ -5,14 +5,14 @@
     android:viewportHeight="19">
   <path
       android:pathData="M19,9.32C19,3.83 14.74,0.1 10,0.1C5.31,0.1 1,3.75 1,9.38C0.4,9.72 0,10.36 0,11.1V13.1C0,14.2 0.9,15.1 2,15.1H3V9C3,5.13 6.13,2 10,2C13.87,2 17,5.13 17,9V16.1H9V18.1H17C18.1,18.1 19,17.2 19,16.1V14.88C19.59,14.57 20,13.96 20,13.24V10.94C20,10.24 19.59,9.63 19,9.32Z"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
   <path
       android:pathData="M7,11.1C7.552,11.1 8,10.652 8,10.1C8,9.548 7.552,9.1 7,9.1C6.448,9.1 6,9.548 6,10.1C6,10.652 6.448,11.1 7,11.1Z"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
   <path
       android:pathData="M13,11.1C13.552,11.1 14,10.652 14,10.1C14,9.548 13.552,9.1 13,9.1C12.448,9.1 12,9.548 12,10.1C12,10.652 12.448,11.1 13,11.1Z"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
   <path
       android:pathData="M16,8.13C15.52,5.28 13.04,3.1 10.05,3.1C7.02,3.1 3.76,5.61 4.02,9.55C6.49,8.54 8.35,6.34 8.88,3.66C10.19,6.29 12.88,8.1 16,8.13Z"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
 </vector>
