+++ app/src/main/res/drawable/ic_vector.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="12">
   <path
       android:pathData="M12,1.209L10.791,0L6,4.791L1.209,0L0,1.209L4.791,6L0,10.791L1.209,12L6,7.209L10.791,12L12,10.791L7.209,6L12,1.209Z"
-      android:fillColor="#AEAFB4"/>
+      android:fillColor="@color/dark_grey"/>
 </vector>
