+++ app/src/main/res/drawable/placeholder.xml
@@ -8,6 +8,6 @@
         android:pathData="M0,0h45v45h-45z"/>
     <path
         android:pathData="M8.606,12.919C9.919,11.587 11.231,10.387 11.813,10.631C12.75,11.006 11.813,12.563 11.25,13.481C10.781,14.269 5.888,20.775 5.888,25.313C5.888,27.712 6.788,29.7 8.4,30.9C9.806,31.95 11.663,32.269 13.35,31.763C15.356,31.181 17.006,29.138 19.087,26.569C21.356,23.775 24.394,20.119 26.737,20.119C29.794,20.119 29.831,22.013 30.038,23.475C22.95,24.675 19.95,30.356 19.95,33.544C19.95,36.731 22.65,39.338 25.969,39.338C29.025,39.338 34.013,36.844 34.763,27.9H39.375V23.212H34.744C34.463,20.119 32.7,15.337 27.188,15.337C22.969,15.337 19.35,18.919 17.925,20.663C16.837,22.031 14.063,25.313 13.631,25.763C13.163,26.325 12.356,27.337 11.55,27.337C10.706,27.337 10.2,25.781 10.875,23.737C11.531,21.694 13.5,18.375 14.344,17.138C15.806,15 16.781,13.538 16.781,10.988C16.781,6.919 13.706,5.625 12.075,5.625C9.6,5.625 7.444,7.5 6.975,7.969C6.3,8.644 5.738,9.206 5.325,9.712L8.606,12.919ZM26.025,34.781C25.444,34.781 24.638,34.294 24.638,33.431C24.638,32.306 26.006,29.306 30.019,28.256C29.456,33.3 27.337,34.781 26.025,34.781Z"
-        android:fillColor="#AEAFB4"/>
+        android:fillColor="@color/dark_grey"/>
   </group>
 </vector>
